Start OBJ parser; Optimize glTF parsing;
