+++ src/headset/openvr.c
@@ -403,9 +403,14 @@ static ModelData* openvrControllerNewModelData(Controller* controller) {
   ModelData* model = lovrAlloc(ModelData);
   size_t vertexSize = sizeof(RenderModel_Vertex_t);
 
-  // Buffers
   model->bufferCount = 2;
-  model->buffers = calloc(model->bufferCount, sizeof(ModelBuffer));
+  model->attributeCount = 4;
+  model->imageCount = 1;
+  model->textureCount = 1;
+  model->materialCount = 1;
+  model->primitiveCount = 1;
+  model->nodeCount = 1;
+  lovrModelDataAllocate(model);
 
   model->buffers[0] = (ModelBuffer) {
     .data = (char*) vrModel->rVertexData,
@@ -418,10 +423,6 @@ static ModelData* openvrControllerNewModelData(Controller* controller) {
     .size = vrModel->unTriangleCount * 3 * sizeof(uint16_t),
     .stride = sizeof(uint16_t)
   };
-
-  // Attributes
-  model->attributeCount = 4;
-  model->attributes = calloc(model->attributeCount, sizeof(ModelAttribute));
 
   model->attributes[0] = (ModelAttribute) {
     .buffer = 0,
@@ -455,32 +456,20 @@ static ModelData* openvrControllerNewModelData(Controller* controller) {
     .components = 1
   };
 
-  // Images
   RenderModel_TextureMap_t* vrTexture = state.deviceTextures[id];
-  model->imageCount = 1;
-  model->images = calloc(model->imageCount, sizeof(TextureData**));
   model->images[0] = lovrTextureDataCreate(vrTexture->unWidth, vrTexture->unHeight, 0, FORMAT_RGBA);
   memcpy(model->images[0]->blob.data, vrTexture->rubTextureMapData, vrTexture->unWidth * vrTexture->unHeight * 4);
 
-  // Textures
-  model->textureCount = 1;
-  model->textures = calloc(model->textureCount, sizeof(ModelTexture));
   model->textures[0] = (ModelTexture) {
     .imageIndex = 0,
     .filter = lovrGraphicsGetDefaultFilter()
   };
 
-  // Material
-  model->materialCount = 1;
-  model->materials = calloc(model->materialCount, sizeof(ModelMaterial));
   model->materials[0] = (ModelMaterial) {
     .colors[COLOR_DIFFUSE] = { 1.f, 1.f, 1.f, 1.f },
     .textures[TEXTURE_DIFFUSE] = 0
   };
 
-  // Primitives
-  model->primitiveCount = 1;
-  model->primitives = calloc(model->primitiveCount, sizeof(ModelPrimitive));
   model->primitives[0] = (ModelPrimitive) {
     .mode = DRAW_TRIANGLES,
     .attributes = {
@@ -492,16 +481,11 @@ static ModelData* openvrControllerNewModelData(Controller* controller) {
     .material = 0
   };
 
-  // Nodes
-  model->nodeCount = 1;
-  model->nodes = calloc(model->nodeCount, sizeof(ModelNode));
   model->nodes[0] = (ModelNode) {
     .transform = MAT4_IDENTITY,
     .primitiveIndex = 0,
     .primitiveCount = 1
   };
-
-  model->rootNode = 0;
 
   return model;
 }
